refactor: MainDetail 리팩토링
